Swap to knex, adds transaction / locks
There was a race condition. If two people were to try to join the same
pair at the same time, it was possible for both pairs to be processed
separately, unaware of the other. One resolution to this problem was
using the SQL `FOR UPDATE` option during our select, requiring the use
of a transaction. Massive didn't support transactions, or passing "for
update", so I opted to replace it with knex.

As an upside, knex is built with promises, so it works more naturally
with what we've currently used to handle our async issues.